requirements.txt
fastapi
pymongo
python-dotenv
uvicorn[standard]
python-multipart
pandas